Preserve JS Object types in state persistance
IndexedDB uses the structured clone algorithm[1] which understands
JS object types like `Date` or `Set` that is lost casting the state
back and forth to JSON strings.

Reactivity is removed from the object being stored which
is why cloneDeep is included as indexedDB requires plain objects.
Removing the Proxy[2] on this.state with lodash cloneDeep.

1: https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm

2: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy